requirements.txt
numpy>=1.22
opencv-python>=4.7.0.68
python-dotenv>=0.21.1
pytorch-lightning>=2.0.0
PyYAML>=6.0
torch>=1.12.1
torchmetrics>=1.0.0
torchvision>=0.13.1
tensorboard>=2.10.1
rich
s3fs
luxonis-ml[all] @ git+https://github.com/luxonis/luxonis-ml.git@babda4b
# tunning related packages
optuna>=3.2.0
psycopg2-binary
# export related packages (only base ones)
onnx>=1.12.0
onnxsim>=0.4.10
onnxruntime>=1.13.1
pycocotools>=2.0.7
blobconverter>=1.4.2
